Document Arista Port-Channel numbers
Airstream is setup on Po8, so it seems the plan is to match the existing
Juniper ae interface numbers.  I updated the website and the Google Doc
spreadsheet accordingly.

I also allocated Po16 to Limelight, as they have two ports allocated.
The welcome message for them on MICE-DISCUSS specifically said they
would be doing a LAG on the new switch.