Add WiscNet remote switch
I have updated the diagram.

I have removed the Northern Lights GigaPoP from the list of remote
switches on the order form.  Once WiscNet's participant port is moved to
their new remote switch, the GigaPoP is going to stop being a remote
switch and will just be an ordinary participant.