Adjust GigaPoP MAC limit
They have 2 IPs, which I believe they have said are on two VRFs on the
same router.  So to allow them to switch routers, they need a limit of
4 MAC addresses.  There was also another MAC, so we set their limit to
5. If that other MAC address disappears (it might have been WiscNet),
then we can reduce them to 4.  Once they get rid of the second IP, they
should be able to be reduced to the normal limit of 1.