Update bootstrap
This upgrades us from bootstrap 3.0.3 and jQuery 1.10.2 to bootstrap
3.4.1 and jQuery 1.12.4.  That was as far as I could go without
breaking the site.

I moved jQuery into the web tree to match bootstrap.  Maybe we should
load these from a CDN.  I don't know.

The initial-scale=1.0 to initial-scale=1 was just to match the example.
The ".0" seems unnecessary.

I dropped the Open Sans font.  The use of it was seemingly inconsistent
on the existing site and I don't feel strongly about the font choice.
That eliminates one more thing to worry about.